Cleanup: Use short for properties editor tab list
Because the active context is stored as a short in DNA it's better
to use short in the code surrounding it. Also adjusted a comment
that reflected an incorrect assumption.

+++ source/blender/editors/include/ED_buttons.h
@@ -28,7 +28,7 @@ extern "C" {
 
 struct SpaceProperties;
 
-int ED_buttons_tabs_list(struct SpaceProperties *sbuts, int *context_tabs_array);
+int ED_buttons_tabs_list(struct SpaceProperties *sbuts, short *context_tabs_array);
 
 #ifdef __cplusplus
 }
